renamed social links file to follow naming convention
